libds: Improve length calculation
It turns out that the length values consists of two subvalues.
The first acts as some sort of base value which determines the
length of the next value, with the sum of both values plus one
forming the final length value.
This should be equivalent to the offset values we where using
earlier, except being faster and easier to understand.

Signed-off-by: Armin Wolf <W_Armin@gmx.de>